refactor: Adapt to frontend port change to 3000
Change cross origin port for react frontend to 3000.
Move wrappers into separate wrappers/ directory.
