+++ spring-backend/src/main/java/practice/recipebase/misc/UploadedRecipe.java
@@ -1,5 +1,7 @@
 package practice.recipebase.misc;
 
+import practice.recipebase.wrappers.UploadedRecipeWrapper;
+
 import java.io.IOException;
 import java.util.List;
