+++ spring-backend/src/main/java/practice/recipebase/wrappers/IngredientsWrapper.java
@@ -1,14 +1,16 @@
-package practice.recipebase.misc;
+package practice.recipebase.wrappers;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.util.List;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class RecipeSiteWrapper {
-    private String recipeSite;
+public class IngredientsWrapper {
+    List<String> ingredientNames;
 }
