+++ spring-backend/src/main/java/practice/recipebase/wrappers/UploadedRecipeWrapper.java
@@ -1,4 +1,4 @@
-package practice.recipebase.misc;
+package practice.recipebase.wrappers;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
